Updates and replacing old libraries - Replace jscoverage with Istanbul - Remove coveralls and serve - Update should, UMD
