+++ .gitignore
@@ -1,5 +1,4 @@
 .DS_Store
 ._*
 node_modules
-lib-cov
-coverage.html
+.nyc_output
