+++ .travis.yml
@@ -1,6 +1,5 @@
 language: node_js
 node_js:
-  - "0.10"
+  - "10"
 after_success:
-  - ./node_modules/.bin/jscoverage lib lib-cov
-  - MS_COV=1 ./node_modules/.bin/mocha test -R mocha-lcov-reporter | ./node_modules/coveralls/bin/coveralls.js
+  - ./node_modules/.bin/mocha test -R mocha-lcov-reporter
