+++ index.js
@@ -1,3 +1,1 @@
-module.exports = process.env.MS_COV
-  ? require('./lib-cov/ms')
-  : require('./lib/ms');
+module.exports = require("./lib/to-ms");
