+++ package.json
@@ -7,10 +7,8 @@
     "test": "tests"
   },
   "scripts": {
-    "test": "mocha",
-    "test-cov": "jscoverage lib lib-cov; MS_COV=1 ./node_modules/.bin/mocha -R html-cov > coverage.html",
-    "browser": "serve test",
-    "make": "umd toMs -c < ./lib/ms.js > dest/to-ms.js"
+    "test": "nyc mocha",
+    "make": "umd toMs -c < ./lib/to-ms.js > dist/to-ms.js"
   },
   "repository": {
     "type": "git",
@@ -35,12 +33,10 @@
   },
   "homepage": "https://github.com/shuvalov-anton/to-ms",
   "devDependencies": {
-    "coveralls": "^2.11.2",
-    "jscoverage": "^0.5.6",
-    "mocha": "^1.21.4",
-    "mocha-lcov-reporter": "0.0.1",
-    "serve": "^1.4.0",
-    "should": "^4.0.4",
-    "umd": "^2.1.0"
+    "mocha": "^6.1.4",
+    "mocha-lcov-reporter": "1.3.0",
+    "nyc": "^14.1.1",
+    "should": "^13.2.3",
+    "umd": "^3.0.3"
   }
 }
