--- test/index.html
@@ -1,18 +1,0 @@
-<html>
-  <head>
-    <title>to-ms - tests</title>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-    <link rel="stylesheet" href="https://cdn.rawgit.com/visionmedia/mocha/1.20.1/mocha.css" />
-  </head>
-  <body>
-    <div id="mocha"></div>
-    <script src="https://cdn.rawgit.com/visionmedia/mocha/1.20.1/mocha.js"></script>
-    <script src="https://cdn.rawgit.com/shouldjs/should.js/master/should.min.js"></script>
-    <script>mocha.setup('bdd')</script>
-    <script src="to-ms.js"></script>
-    <script src="test.js"></script>
-    <script>
-      mocha.run();
-    </script>
-  </body>
-</html>
